Classes to call the API.
There are two main things here. The code to call the graphql api which
is pretty standard and is doing the same thing as the backend checks.

The other is using sttp to call keycloak to get a token for the
temporary user using the temporary user credentials with a password
grant.

This means we can create consignments and update transfer agreements
without having to go through the preceeding steps which should speed up
the tests a bit, especially when we get to later pages.