Fix redirect step
Redirect step was checking for the page name in the page title. The upload page title is "Upload Records" therefore the test was passing even though no redirect took place, as was checking for the word "Record" in the page title.

Changed to check the URL to ensure unique to each page