Increase upload file size
"Progress bar visible" test intermittently failing. Caused by the file upload completing, and page redirect occuring before test completed check for progress bar element. The element was not present on the redirect page, causing the condition to timeout, and the test to fail

Increased the upload file size for the "Progress bar visible" to 10mb. With this size file the test passed 10 times in row during testing