Reconfigure Keycloak admin client to fix token expiry bug
The tests have started to fail on Jenkins because the access token used
to create and delete users is expiring. The token expires after five
minutes, and the tests now take longer than five minutes to run because
some of them have to wait for the file checks to finish.

Before, we created the Keycloak admin client using an access token, which
is the one that expires. Now we create the client using the client ID and
secret. The client can now store a refresh token, so it can fetch a new
access token when the original one expires.