Change the way we check if the previous build failed.
We were using fixed in the post block but this would fire if the
previous test run was for staging which failed and then the next test
run for integration started working which isn't what we want.

This now moves recursively through the previous builds until it finds
the previous one with the same stage as the current build and then
returns that status.