Change the tests to work with the new file checks page.
The file checks page no longer redirects to the results page when the
checks complete. Instead, it shows or hides the notification banner
based on the checks status.
I've changed the tests to check for this. I also moved out some tests
that really belonged in the records results feature page. I've also
updated the wording we're checking for on some of these.