Add timeout for checking that file check link has been enabled
This step was failing on staging. We think it might be because
Cucumber is checking the button state as soon as the banner appears,
but it might be running so fast that it catches the page in the
intermediate state when the button hasn't been enabled yet.

This change will force it to wait 2 seconds before failing.