Increase timeout on upload page.
The AWS.config.credentials are now set after the call to the cookies
endpoint. This endpoint can be quite slow, sometimes taking more than 10
seconds if it has to cold start so I've increased the timeout here to
try to stop this failing.

This step passed on another test in the same run so it does work, it
just times out sometimes.