Upgrade everything to allow us to use dev tools.
In selenium 4 there is HasDevTools interface which allows you to log all
network requests.
The version of selenium bundled with scalatestplus-play was 3.x.x so
I've got rid of this dependency and added the selenium and typesafe ones
individually. I've also updated Cucumber because why not.
There are a few code changes to migrate over and some general tidying
up.
I've added in a section to the Before function to initialise the network
logging.