Setup test users with keycloak groups
The user setup for the tests was written before Keycloak groups were introduced.

Test users are now assigned to the relevant user group(s). This is how real users are setup.

The user groups provide the individual users with the necessary permissions and attributes.

These changes will also help when Keycloak groups are introduced to support court judgment transfers.