Add wait before upload.
We used to check if the AWS config object was populated but that has
been removed in the newer version of the sdk so there is now no way to
determin if the cookies endpoint has responded without changing the
frontend just to make the tests work.

So we're just sleeping for 15 seconds. We can always tweak this later.