Run everything in one container.
I've changed the transfer-frontend container so it can run
ssm_get_parameter so we don't have to wait for the aws container to
start. This has saved 3-4 minutes.

I had to add the INTG_ACCOUNT and STAGING_ACCOUNT parameters to make
this work which means that the tests can now use those instead of the
INTG_AWS_ACCOUNT_NUMBER variable.

This can't be merged until
https://github.com/nationalarchives/tdr-jenkins/pull/275 is in and
deployed.