Run tests in parallel in multiple containers.
I have tried everything I can to get this to work dynamically by reading
the feature files and running them in parallel on a loop and I don't
think it can be done.
The issue comes down to the fact that if you want a dynamic loop, you
need to call it from within a script tag and you can't have an agent
block within a script tag.
So the only way to do it is like this with the stages set explicitly.
There's a lot of duplication and if there are ever new feature files
then we'll have to add it in here manually so I'm not sure.