Put the concurrency back in.
We're hitting the limit on the number of rules a security group can have
because these tests are running at the same time on the same
environment.

This should only allow one set of tests to run per environment now. If
this starts taking for ever with the pull request tests then we can
rethink it but this is the only thing I can think of to stop them
failing all the time now.