Add tests for the download metadata page
These are the standard tests checking for page elements and making sure
only logged in standard users can see it but there's also a test to
check the output of the CSV file.

At the moment, all of the "AllowExport" columns in the "FileProperty"
table are set to false so there is only one column in the export for the
file name.

There is a ticket
https://national-archives.atlassian.net/browse/TDR-2382 to update these.

This PR can't be merged until
https://github.com/nationalarchives/tdr-transfer-frontend/pull/2186 is
merged and that's waiting on
https://github.com/nationalarchives/tdr-generated-graphql/pull/672 in
turn.

If TDR-2382 is merged before the other two PRs are deployed then I'll
update these tests but if not, I'll merge them as is and we can update
them once TDR-2186 is merged.