Fix the file check results tests.
Some of the tests were testing to see if we ended up on the additional
metadata page but if you go there without setting the Upload consignment
status to completed, it bounces you back to the upload page so this
wasn't working.

I set the consignment status but then it was complaining about a missing
parent folder. I've honestly no idea how this ever worked because the
code to check for that is a year old but there you are, it seems ok now.